Refactor collision id helpers

+++ gg2/structs/collision/CollisionPair.h
@@ -4,6 +4,6 @@
 
 struct CollisionPair
 {
-    uint16_t a[MAX_COLLISION_PAIRS];
-    uint16_t b[MAX_COLLISION_PAIRS];
+    uint32_t a[MAX_COLLISION_PAIRS];
+    uint32_t b[MAX_COLLISION_PAIRS];
 };
